requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
PyMySQL==1.1.0
Werkzeug==2.3.7
mysql-connector-python==8.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
